--- vodomjer_web/src/assets/dragdropframe.css
@@ -1,26 +1,0 @@
-.frame-container {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.frame {
-  border: 1.5px solid #000;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.button-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: 565px;
-  margin-left: -470px;  
-}
-
-.button {
-  border: 1px solid #000;
-  padding: 10px 20px; 
-}
