--- vodomjer_web/src/assets/navbar.css
@@ -1,26 +1,0 @@
-.navbar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    background-color: #8DECB4;
-    color: #141E46;
-    padding: 10px 20px;
-    border: 1px solid black;
-  }
-  
-  .logo img {
-    width: 50px;
-    height: auto;
-  }
-  
-  .title {
-    flex-grow: 1; 
-    text-align: center;
-  }
-  
-  .links a {
-    color: #141E46;
-    text-decoration: none;
-    margin-left: 20px;
-  }
-
