--- vodomjer_web/src/assets/sidemenu.css
@@ -1,29 +1,0 @@
-.side-menu {
-    display: flex;
-    border: 1px solid black;
-    width: 250px;
-    height: 100vh;
-    background-color: #8DECB4;
-    color: #141E46;
-  }
-  
-  .side-menu ul {
-    flex-grow: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    align-items: center;
-    list-style-type: none;
-    padding: 0;
-  }
-  
-  .side-menu li {
-    padding: 0;
-  }
-  
-  .side-menu a {
-    color: #141E46;
-    font-size: 1.3rem;
-    text-decoration: none;
-  }
-
