--- vodomjer_web/src/assets/uploadframe.css
@@ -1,19 +1,0 @@
-.frame-container {
-    position: relative;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-left: 300px;
-    margin-right: 350px;
-    margin-bottom: 80px;
-  }
-  
-  .frame2 {
-    border: 1.5px solid #000;
-    display: flex;
-    justify-content: center;
-    align-items: flex-start;
-    text-align: center;
-    font-size: 1.2rem;
-    padding-top: 15px;
-  }
