--- vodomjer_web/src/custom.d.ts
@@ -1,6 +1,0 @@
-// src/custom.d.ts
-declare module '*.png' {
-    const value: any;
-    export default value;
-  }
-
