+++ web_stranica/vodomjer_web/package.json
@@ -10,7 +10,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.23.1",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -42,5 +42,4 @@
     "@types/react-router-dom": "^5.3.3"
   },
   "proxy": "http://localhost:5000"
-
 }
