+++ web_stranica/vodomjer_web/src/components/Sidemenu.js
@@ -5,7 +5,7 @@ const Sidemenu = () => {
   return (
     <div className="side-menu">
       <ul>
-        <li><a href="/">Učitavanje datoteka vodomjera/bankovnih izvoda</a></li>
+        <li><a href="/ucitavanje">Učitavanje datoteka vodomjera/bankovnih izvoda</a></li>
         <li><a href="/izvjestaj">Formiranje izvještaja</a></li>
         <li><a href="/slanje">Slanje izvještaja</a></li>
         <li><a href="/baza">Uvid u bazu podataka</a></li>
