+++ web_stranica/vodomjer_web/src/components/UploadFrame.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import '../assets/uploadframe.css';
+
+const Upload = ({ width, height }) => {
+  return (
+    <div className="frame-container">
+      <div className="frame2" style={{ width, height }}>
+        UPLOADANE DATOTEKE
+      </div>
+    </div>
+  );
+}
+
+export default Upload;
